The MyAwesomePackage Module

SmallCouplingDynamicCavity.SmallCouplingDynamicCavity — Module

Module for "SmallCouplingDynamicCavity.jl" – A package for a Small Coupling expansion of the Dynamic Cavity method for epidemic inference.

Module Index

SmallCouplingDynamicCavity.SI

SmallCouplingDynamicCavity.SIR

SmallCouplingDynamicCavity.SIRS

SmallCouplingDynamicCavity.SIS

SmallCouplingDynamicCavity.run_SCDC

SmallCouplingDynamicCavity.sim_epidemics

SmallCouplingDynamicCavity.sim_epidemics

SmallCouplingDynamicCavity.sim_epidemics

SmallCouplingDynamicCavity.sim_epidemics

Detailed API

SmallCouplingDynamicCavity.SI — Method

SI(
    εᵢᵗ::Union{Float64,Array{Float64,2}},
    NV::Int,
    T::Int)

Defines the SIS infection model.

Arguments

εᵢᵗ: Self-infection probability. Can be either a Float64 (constant over all nodes and times) or a NVxT matrix.

NV: Number of nodes of the contact graph.

T: Number of time-steps.

SmallCouplingDynamicCavity.SIR — Method

SIR(
    εᵢᵗ::Union{Float64,Array{Float64,2}},
    rᵢᵗ::Union{Float64,Array{Float64,2}},
    NV::Int,
    T::Int)

Defines the SIS infection model.

Arguments

εᵢᵗ: Self-infection probability. Can be either a Float64 (constant over all nodes and times) or a NVxT matrix.

rᵢᵗ: Recovery probability. Can be either a Float64 (constant over all nodes and times) or a NVxT matrix.

NV: Number of nodes of the contact graph.

T: Number of time-steps.

SmallCouplingDynamicCavity.SIRS — Method

SIRS(
    εᵢᵗ::Union{Float64,Array{Float64,2}},
    rᵢᵗ::Union{Float64,Array{Float64,2}},
    σᵢᵗ::Union{Float64,Array{Float64,2}},
    NV::Int,
    T::Int)

Defines the SIS infection model.

Arguments

εᵢᵗ: Self-infection probability. Can be either a Float64 (constant over all nodes and times) or a NVxT matrix.

rᵢᵗ: Recovery probability. Can be either a Float64 (constant over all nodes and times) or a NVxT matrix.

σᵢᵗ: Loss of immunity probability. Can be either a Float64 (constant over all nodes and times) or a NVxT matrix.

NV: Number of nodes of the contact graph.

T: Number of time-steps.

SmallCouplingDynamicCavity.SIS — Method

SIS(
    εᵢᵗ::Union{Float64,Array{Float64,2}},
    rᵢᵗ::Union{Float64,Array{Float64,2}},
    NV::Int,
    T::Int)

Defines the SIS infection model.

Arguments

εᵢᵗ: Self-infection probability. Can be either a Float64 (constant over all nodes and times) or a NVxT matrix.

rᵢᵗ: Recovery probability. Can be either a Float64 (constant over all nodes and times) or a NVxT matrix.

NV: Number of nodes of the contact graph.

T: Number of time-steps.

SmallCouplingDynamicCavity.run_SCDC — Method

run_SCDC(
    model::EpidemicModel{TI,TG},
    obsprob::Function,
    γ::Float64,
    maxiter::Int64,
    epsconv::Float64,
    damp::Float64;
    μ_cutoff::Float64 = -Inf,
    callback::Function=(x...) -> nothing) where {TI<:InfectionModel,TG<:Union{<:AbstractGraph,Vector{<:AbstractGraph}}}

Runs the SCDC inference.

Arguments

model: Epidemic model.

obsprob: Evidence (likelihood) probability p(O|x) of an obseravtion O given the planted state x.

γ: Probability of being a patient zero.

maxiter: Maximum number of iterations.

epsconv: Convergence threshold of the algorithm.

damp: Damping factor of the algorithm.

μ_cutoff: Lower cut-off for the values of μ.

callback: Callback function.

SmallCouplingDynamicCavity.sim_epidemics — Method

sim_epidemics(
    model::EpidemicModel{SI,TG};
    patient_zero::Union{Vector{Int},Nothing}=nothing,
    γ::Union{Float64,Nothing}=nothing) where {TG<:Union{<:AbstractGraph,Vector{<:AbstractGraph}}}

Simulates the epidemic outbreak given a SI model.

Arguments

model: The SI epidemic model.

patient_zero: Vector of patients zero. Default is "nothing", meaning that the patients zero are chosen at random with probability γ.

γ: Probability of being a patient zero. Default is "nothing", meaning that it is fixed to0 1/NV, where NV is the number of nodes of the contact graph.

SmallCouplingDynamicCavity.sim_epidemics — Method

sim_epidemics(
    model::EpidemicModel{SIR,TG};
    patient_zero::Union{Vector{Int},Nothing}=nothing,
    γ::Union{Float64,Nothing}=nothing) where {TG<:Union{<:AbstractGraph,Vector{<:AbstractGraph}}}

Simulates the epidemic outbreak given a SIR model.

Arguments

model: The SIR epidemic model.

patient_zero: Vector of patients zero. Default is "nothing", meaning that the patients zero are chosen at random with probability γ.

γ: Probability of being a patient zero. Default is "nothing", meaning that it is fixed to0 1/NV, where NV is the number of nodes of the contact graph.

SmallCouplingDynamicCavity.sim_epidemics — Method

sim_epidemics(
    model::EpidemicModel{SIRS,TG};
    patient_zero::Union{Vector{Int},Nothing}=nothing,
    γ::Union{Float64,Nothing}=nothing) where {TG<:Union{<:AbstractGraph,Vector{<:AbstractGraph}}}

Simulates the epidemic outbreak given a SIRS model.

Arguments

model: The SIRS epidemic model.

patient_zero: Vector of patients zero. Default is "nothing", meaning that the patients zero are chosen at random with probability γ.

γ: Probability of being a patient zero. Default is "nothing", meaning that it is fixed to0 1/NV, where NV is the number of nodes of the contact graph.

SmallCouplingDynamicCavity.sim_epidemics — Method

sim_epidemics(
    model::EpidemicModel{SIS,TG};
    patient_zero::Union{Vector{Int},Nothing}=nothing,
    γ::Union{Float64,Nothing}=nothing) where {TG<:Union{<:AbstractGraph,Vector{<:AbstractGraph}}}

Simulates the epidemic outbreak given a SIS model.

Arguments

model: The SIS epidemic model.

patient_zero: Vector of patients zero. Default is "nothing", meaning that the patients zero are chosen at random with probability γ.

γ: Probability of being a patient zero. Default is "nothing", meaning that it is fixed to0 1/NV, where NV is the number of nodes of the contact graph.